implemented payments

+++ init/mysql/initdb.d/billing.sql
@@ -63,18 +63,18 @@ CREATE TABLE orders (
     PRIMARY KEY (id)
 );
 CREATE TABLE order_items (
-    id          BIGINT  NOT NULL AUTO_INCREMENT,
-    price_id    BIGINT  NOT NULL,
+    id           BIGINT  NOT NULL AUTO_INCREMENT,
+    price_id     BIGINT  NOT NULL,
     orders_items BIGINT,
-    quantity    TINYINT UNSIGNED DEFAULT 1,
+    quantity     TINYINT UNSIGNED DEFAULT 1,
     FOREIGN KEY (orders_items) REFERENCES orders (id),
     PRIMARY KEY (id)
 );
 CREATE TABLE order_txs (
-    id          BIGINT      NOT NULL AUTO_INCREMENT,
-    status      CHAR(8)     NOT NULL,
-    orders_txs  BIGINT      NULL,
-    created_at  DATETIME(3) NOT NULL,
+    id          BIGINT       NOT NULL AUTO_INCREMENT,
+    status      CHAR(8)      NOT NULL,
+    orders_txs  BIGINT       NULL,
+    created_at  DATETIME(3)  NOT NULL,
     FOREIGN KEY (orders_txs) REFERENCES orders (id),
     PRIMARY KEY (id)
 );
@@ -82,7 +82,7 @@ CREATE TABLE order_txs (
 # |payment_instruments|
 #  -------------------
 CREATE TABLE payment_instruments (
-     id           BIGINT      NOT NULL AUTO_INCREMENT,
+     id           TINYINT     NOT NULL AUTO_INCREMENT,
      type         CHAR(12)    UNIQUE NOT NULL,
      type_i18n_id BIGINT      NULL,                # reserved for the future
      PRIMARY KEY (id)
@@ -93,7 +93,7 @@ CREATE TABLE payment_instruments (
 CREATE TABLE payments (
      id                                                     BIGINT NOT NULL AUTO_INCREMENT,
      orders_payments                                        BIGINT UNIQUE,
-     instrument_payments                                    BIGINT,
+     instrument_payments                                    TINYINT,
      FOREIGN KEY (orders_payments)                          REFERENCES orders (id),
      FOREIGN KEY (instrument_payments)                      REFERENCES payment_instruments (id),
      PRIMARY KEY (id)
